+++ library/unified/server/source/unified_compressor.cpp
@@ -121,13 +121,13 @@ void sirius::library::unified::compressor::after_video_compressing_callback(int3
 		_front->after_video_compressing_callback(index, compressed, size, before_compress_timestamp, after_compress_timestamp);
 }
 
-void sirius::library::unified::compressor::after_video_compressing_callback(int32_t count, int32_t * x, int32_t * y, int32_t * width, int32_t * height, uint8_t ** compressed, int32_t * size, long long before_compress_timestamp, long long after_compress_timestamp)
+void sirius::library::unified::compressor::after_video_compressing_callback(int32_t count, int16_t * x, int16_t * y, int16_t * width, int16_t * height, uint8_t ** compressed, int32_t * size, long long before_compress_timestamp, long long after_compress_timestamp)
 {
 	if (_front)
 		_front->after_video_compressing_callback(count, x, y, width, height, compressed, size, before_compress_timestamp, after_compress_timestamp);
 }
 
-void sirius::library::unified::compressor::after_video_compressing_callback(int32_t x, int32_t y, int32_t width, int32_t height, uint8_t * compressed, int32_t size, long long before_compress_timestamp, long long after_compress_timestamp)
+void sirius::library::unified::compressor::after_video_compressing_callback(int16_t x, int16_t y, int16_t width, int16_t height, uint8_t * compressed, int32_t size, long long before_compress_timestamp, long long after_compress_timestamp)
 {
 	if (_front)
 		_front->after_video_compressing_callback(x, y, width, height, compressed, size, before_compress_timestamp, after_compress_timestamp);
